% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_path_finder.R
\name{ima_path_finder}
\alias{ima_path_finder}
\title{Title}
\usage{
ima_path_finder(image_path = "", api_key = api_key,
  api_secret = api_secret, ...)
}
\arguments{
\item{image_path}{Where is the image url or local path}

\item{api_key}{Api Key}

\item{api_secret}{Api Secret}

\item{...}{More arguments}
}
\value{
The url complete where to call the IMAGGA API
}
\description{
Title
}
\examples{

}
