% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_rescale.R
\name{ima_rescale}
\alias{ima_rescale}
\title{Rescale big images if its needed}
\usage{
ima_rescale(image_path = "", ...)
}
\arguments{
\item{image_path}{File path where the image is located}

\item{...}{More arguments passed to this function}
}
\value{
Path directory where te rescale image is.
}
\description{
Rescale big images if its needed
}
\examples{

}
