% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_tag.R
\name{ima_tag}
\alias{ima_tag}
\title{A function to call the tagger option in IMAGA API}
\usage{
ima_tag(image_path = "", api_key = api_key, api_secret = api_secret,
  limit_tag = -1, syssleep = 5, rescale = FALSE)
}
\arguments{
\item{image_path}{The path where the image is located, it can be a url or a local directory.}

\item{api_key}{Api Key}

\item{api_secret}{Api Secret}

\item{limit_tag}{A number to limit the total n of tags to receive.}

\item{syssleep}{Modify system sleep. By default 5 to be polite.}

\item{rescale}{If your image is too big you can rescale it.}
}
\value{
Return a data frame containing \code{limit_tag} from an image.
}
\description{
A function to call the tagger option in IMAGA API
}
\examples{

}
