% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ima_upload.R
\name{ima_upload}
\alias{ima_upload}
\title{Upload an image and return upload_id}
\usage{
ima_upload(image_path = "", api_key = api_key,
  api_secret = api_secret, ...)
}
\arguments{
\item{image_path}{Image path}

\item{api_key}{Api Key}

\item{api_secret}{Api Secret}

\item{...}{More arguments to pass}
}
\value{
This function returns \code{upload_id}
}
\description{
Upload an image and return upload_id
}
\examples{

}
