% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaggar-package.R
\docType{package}
\name{imaggar-package}
\alias{imaggar}
\alias{imaggar-package}
\title{imaggar: R client for accessing IMAGGA’s REST API.}
\description{
R client for accessing IMAGGA’s REST API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VictorSuarezL/imaggar}
  \item Report bugs at \url{https://github.com/VictorSuarezL/imaggar/issues}
}

}
\author{
\strong{Maintainer}: Victor Sanz \email{victor.sanz@uca.es} (YOUR-ORCID-ID)

}
\keyword{internal}
